s32g: Add basic UART console support
Program LINFlexD in UART mode for a basic console support. Only Tx is
supported.

Also, as a performance-enhancing hack, console parameters are heavily
(and illegally!) skewed on the S32G simulator (and only on simulator),
in order to achieve a better console throughput. This is a VDK-only hack
and will be wiped when the real Si is out.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-3452