s32g: Move BL31 after BL33 in the SRAM layout
U-Boot unilaterally assumes the SRAM (at least the start of it) is free
for use and installs the MMU tables there, overwriting our exception
vectors.
Before we properly relocate the BL31 in DDR in order to preserve and
protect the PSCI callbacks, we place BL31 *after* U-Boot in SRAM, where
it will potentially escape overwriting. This has the added advantage of
extra space, allowing us to compile the TF-A with -O0 without
overlapping with adjacent images in SRAM.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-3452