s32g: Do not memset cpu context on warm boot
As a workaround imposed by the fact that we aren't quite powering down
the system yet, but instead are splicing together the suspend and resume
execution paths, we must avoid context zeroization which destroys
registers used by the non-secure world upon resume.

In the U-Boot unit tests, the issue is resolved by simply saving the context
separately, but for Linux that tends to become uglier, so it is simpler
to just not do it at all in the TF-A.

To be revisited once we have full power management support in the
functional simulator.

Issue: ALB-3702

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>