s32g: Exclude s32g_xrdc.c from build
The XRDC driver is becoming ever more obsolete. The latest addition to
the table is the discontinuity of the DDR physical address space,
whereby the first and the last 2GB may not be adjacent. Since we aren't
in fact using XRDC at all, remove it from the build for now.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-4474