s32g274a: Reconfigure Ncore before starting secondary cluster
The secondary CA53 cluster is not configured as an Ncore master at boot.
We mustn't do that before we've taken at least one of A53 cores 2 or 3
out of reset. That comes with a double barrier between the primary core
and the secondary, whereby the latter waits in a 'wfi' holding pen, then
again waits for Ncore to be configured.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-4497