s32g274a: Revert "s32g274a: Define HW_ASSISTED_COHERENCY"
Turns out that HW_ASSISTED_COHERENCY causes the PSCI_CPU_OFF call to
crash because of the following scenario:

psci_do_pwrdown_sequence
    enters with caches enabled
    pushes X30 onto the stack
    calls prepare_cpu_pwr_dwn
	prepare_cpu_pwr_dwn
	    calls cortex_a53_core_pwr_dwn
		cortex_a53_core_pwr_dwn
		    disables dcache on the current core
		    cleans and invalidate L1 cache
		    disables SMPEN bit
		    (...as per the procedure described in the
		    A53 reference manual (DDI0500D), "Individual
		    core shutdown mode")
		    ret
	    ret
	ret
    pops X30 from the stack
    return

Note that at the time when X30 was last pushed onto the stack, caches
were enabled. At the time when it is popped, caches are disabled, but
only L1 has been cleaned. The manual procedure does not specify that the
clean must be to the PoC (Point of Coherency). As such, X30 may not have
been flushed all the way to the main memory, which remained incoherent.
Function psci_do_pwrdown_sequence() will return to an invalid address
and crash.

This reverts commit 4a4c315c70377eb910058a741e9004b9600ecaaf.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-4909