s32g274a: Add RTC interrupt
Route RTC interrupt to A53_0 at EL3 so we wake up from wfi.
Allow the secondary cores to receive the SGI in order to wake from wfi
and proceed all the way to the kernel.

NOTE - Work In Progress: apparently the secondary core's context is not
completely restored, so that it eventually crashes inside the kernel,
in secondary_start_kernel() at the first operation with the stack.
While this issue is being debugged, the boot process of the secondaries
is stopped short inside the TF-A via a branch-to-self.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-4910
Issue: ALB-4909