tf-a: fix boot flow from eMMC
eMMC device type would not get selected because of a typo in the macro
that is guarding this selection. Also, there was no way of selecting
the eMMC boot flow when building tf-a. Hence, the platform makefile
now checks the presence of S32G_BOOT_FROM_EMMC when executing make.

Issue: ALB-7007
Signed-off-by: Bogdan-Gabriel Roman <bogdan-gabriel.roman@nxp.com>