s32gen1:uart: Suppress printing to shared console
Sharing our console with Linux is not a safe way of printing, because of
possible differences in our UART driver vs. whatever Linux context may
use.
Add a compile-time variable in platform.mk and conditionally suppress
printing in BL31. The switch can be turned on if deemed safe from the
UART sharing viewpoint.
This patch should be considered a workaround.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-6945