s32g: add support for reading FIP from a defined memory offset
A new parameter is added in Makefile: FIP_MEM_OFFSET

When it is not defined, the existing procedure for loading FIP is used:
it is read from the QSPI/MMC offsets depending on boot source

When FIP_MEM_OFFSET is defined as a memory address, FIP is loaded from
the defined memory address independent of the boot source. It is the M7
bootloader responsibility to load the fip at the specified address
before starting BL2.
FIP_MEM_OFFSET should be chosen such that it does not overlap with the
existing memory mapping and SRAM usage.

How to build example:
make DEBUG=1 PLAT=s32g2 BL33=<path to u-boot.bin> FIP_MEM_OFFSET=0x34520000

Issue: ALB-7495

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>