s32g: fix accessing s32g_policies array outside array size
This is reproducible when optee is enabled. If ATF is compiled with
DEBUG=1, the issue is detected in an assert:
	ASSERT: plat/nxp/s32/s32g/s32g_storage.c:337:
		image_id < ARRAY_SIZE(s32g_policies)
If compiled without DEBUG, the memory is written outside s32g_policies
array.

The fix consists of correct dimensioning of s32g_policies array if optee
is enabled.

Issue: ALB-8232

Signed-off-by: Catalin Udma <catalin-dan.udma@nxp.com>