bl2: Exclude ECC range from U-Boot DT nodes
Added a ft_fixup function to read each memory node
from U-Boot's DT, check if the ranges are correct based
on the DRAM controller registers and the ECC feature being
enabled or not, and update the ranges with the correct ones.

U-Boot will parse the DT with the memory nodes as modified by
ATF, and no longer run the ECC exclusion algorithm.

Issue: ALB-8325
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>