s32: Introduce FIP_HDR_SIZE makefile variable
FIP_HDR_SIZE is defined to allow standalone build of the BL2 stage.
This is needed because BL2 stage needs to know the size of the FIP
header to determine the address of the device tree and FIP header.

This allows to compile BL2 for RDB2 using the command:
make PLAT=s32g274ardb2 FIP_SD_OFFSET=0x1000 FIP_HDR_SIZE=0x200 bl2

Issue: ALB-8551
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>