scmi: s32: fix CLOCK_GET_ATTRIBUTES response
The response would not correctly return the enabled bit for some clocks.
This commit removes the enabled member from the `struct s32gen1_scmi_clk`
and uses the `clk_states` instead. This means that the message will
return if a clock is enabled for a given agent, not if it is physically
enabled.

Issue: ALB-8770

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>