s32: mmc: Poll SDSTB when changing the clock
The VEND_SPEC_PER_CLKEN and VEND_SPEC_CARD_CLKEN bits are marked as
Reserved in the SoC Reference Manual, therefore we should not touch them.
Indeed, these bits are present and described in the USDHC Block Guide, but
still not mentioned as required to be cleared/set when changing the
frequency.

The only required steps before changing the clock are to poll for the SDSTB
bit and ensure FRC_SD_CLK is 0. Since the later is not used in this driver,
we'll only check the former.

The udelay() call was likely there to wait for the clock to be stable after
being changed, which is an incorrect alternative to polling for SDSTB.

Issue: ALB-9078
Signed-off-by: Dan Nica <dan.nica@nxp.com>