ddr: Add CUSTOM_DDR_DRV make parameter for external driver
Added the handling of CUSTOM_DDR_DRV parameter, which should
point to a folder that contains the DDR code in the state
in which it is generated by DDR Tool.

With the CUSTOM_DDR_DRV passed to the make command, the
"import_ddr_drv" rule should be invoked first, such that
the DDR files are changed and adapted to ATF, and
the compilation command afterwards.

It can be used as following, for example:
make CROSS_COMPILE=... ARCH=aarch64 PLAT=... BL33=...
	CUSTOM_DDR_DRV=<relative_path_to_ddr_code> import_ddr_drv
make CROSS_COMPILE=... ARCH=aarch64 PLAT=... BL33=...
»       CUSTOM_DDR_DRV=<relative_path_to_ddr_code>

Issue: ALB-9210
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>