s32: ddr: Switch to quote include statements
Changed the include statements to be quote-based, in order
to align to the DDR Tool generated format. This means that
the DDR driver folder has to be added to the PLAT_INCLUDES
directly.

Also if external DDR firmware is used, the CUSTOM_DDR_DRV
needs to be added to PLAT_INCLUDES too, before the normal
DDR Driver folder, since the -I<include> folders are searched
for headers in the order they were included, from left to right.

Issue: ALB-9354
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>