s32cc: Disable the core even if its clock seems off
This is useful for the secondary cores of a cluster where
the clock status bit is reserved. It's safe to stop
these cores even if they don't have full control of their
clock. The master core (lowest MIDR ID) of the cluster
should to cut the clock during its shutdown. The SW upper
layers must contain glue logic where the slave cores of the
cluster are shutdown prior to the master core.

Skip turn-off procedure for cores which are in reset.

Issue: ALB-9435
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>