s32: XIP support for BL2 when compiled with FIP_MEMORY_OFFSET flag
Fip.bin binary is deployed directly into memory when FIP_MEMORY_OFFSET
is specified. This means that BL31 is placed right after the BL2 data
section. There’s no reserved section for BL2’s stack and bss in this
case which will affect the BL31 binary once the BL2 initializes and uses
the stack. A padding is added after the BL2 binary as a placeholder for
stack and bss to avoid BL31’s corruption when BL2 is executed in place.

Issue: ALB-9607
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>