s32: ddr: Update to driver/fw from S32CT 1.6 Update 6
This update comes with improvements regarding the DDR PHY impedance
calibration sequence. Also, the constraints on the Refresh Command
were removed, because the default configuration is working on both
Legacy and Modified mode devices, thus, no longer needing a separate
configuration for RDB3 Rev. F, which will be treated in a following commit.

Issue: ALB-9722
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>