dt-bindings: s32cc: clocks: Reassign PFE clock ids
This reassigns ids for PFE clocks to ease a sequence of clocks turn-off
based on ascending order of the clocks. Placing the PFE_PE clock at the
end of the PFE clocks forces the logic to keep the PFE partition and HW
block on until the last moment. BASE_RESET_AGENT_CONFIGURATION is an
SCMI command whose implementation disables clocks by iterating Ids in
ascending order.

Issue: ALB-9888
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>