s32r: fip_mem: Increase 'MAX_XLAT_TABLES' and 'MAX_MMAP_REGIONS'
When booting with FIP in Mem support (XIP),
the 'MAX_XLAT_TABLES' should be increased, for S32R platforms.

As part of this commit, we increase 'MAX_XLAT_TABLES' for S32R
platforms in order to work in all scenarios.

Also, as a leftover from the original FIP in memory (XIP) support,
'MAX_MMAP_REGIONS' should also be increased by value one,
in case of S32R platforms, as it was done for S32G platforms.

Issue: ALB-9909
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>