s32: scmi: Reset the clocks of the SCMI agents before suspend
Disable the SCMI clocks by iterating through all agents and
their clocks left enabled before the platform suspend. This is
a must to avoid SW and HW being out of sync during the resume
operation where the clock driver will have the impression that
a clock is already enabled (based on its metadata), although it
is disabled from the hardware perspective.

Issue: ALB-9888
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>