s32gen1-clock-freq: adjust the qspi clock frequency based on board
EVB Boards have a flash memory which can operate at a frequency of
maximum 166 MHz. However, getting this frequency would require further
modifications in the clock tree configuration. For now, set the
frequency to 133 MHz as a conservative approach, since this frequency
can be obtained with the current clock configurations.

The rest of the boards have flash memories which can operate at a
frequency of 200 MHz.

Issue: ALB-9840
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>