s32: scmi: Change the behaviour of CLOCK_RATE_SET command
The SCMI specification is not very detailed about the CLOCK_RATE_SET
command. It leaves space for interpretation to accommodate multiple
platforms with different restrictions.

On S32CC platforms, it will not change the state of the clock as before
this change. Only the clock driver metadata will be updated if the clock
is disabled before the command is received. For enabled clocks, the
hardware will be reprogrammed to reflect the requested frequency.

Issue: ALB-9934
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>