s32: clk: Add partition link blocks
This new block type has the same role as the previously introduced
‚partition block link’. It helps to keep a reference to a partition for
a clock that does not have a partition block attached to it but depends
on the partition state. For example, the PFE_PE clock from S32G2 can
only be enabled if the PFE partition is taken out of reset.

Issue: ALB-9888
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>