fdts: s32cc: Add node for U-Boot SoC Driver
Added a "chipid" node as suggested by the U-Boot UCLASS_SOC guidelines,
with the "nxp,s32cc-soc" compatible that will be correspondent to the
newly added U-Boot S32CC SoC Driver. Reporting SoC-related info now
is the responsibility of the SoC driver instead of the CPU driver,
so link the relevant nvmem cells to the DT Node.

Issue: ALB-10082
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>