ddr: Override reset function with platform-specific implementation
The newly added `deassert_ddr_reset()` function should be overridden
with a platform-specific implementation that does not directly access
the MC_RGM registers.

Used the existing `reset_ddr_periph()` function for doing that through
the SCMI Reset Domain protocol, and changed its signature to override
the `deassert_ddr_reset()` function.

Also fixed an include for a DDR header which needs to be a quote-based
include statement, in order to make use of the mechanism for using
externally provided DDR code.

Issue: ALB-10131
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>