nxp/s32/clk: Fix divider calculation
Adjust the calculation of the divider value to ensure that the resulting
division factor does not exceed the dividend when the divisor is
multiplied by it. This will prevent assumptions that the parent
frequency is higher than it actually is. Additionally, it may be
necessary to adjust the resulting division factor so that the rate
obtained after division does not fall below the specified frequency.

Issue: ALB-10430
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>