plat/nxp/s32: Restore clock frequency when SCMI set rate failed
Sometimes, the frequency requested may not match the current state of
the system. When this happens, the clock associated with the failed rate
set operation will either stay disabled or set at a fail-safe rate. To
prevent this from happening, we will revert back to the original rate
used prior to the rate-set operation.

Issue: ALB-10430
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>