ERR051700: drivers: refactor get_reset_regs
In order to avoid wrong usage of mmio_write to
MC_RGM_PRSTn[PERIPH_x_RST], we need to encapsulate both read and write
operations. This means the translation of the address must be done in
the newly created functions. Currently this is not possible because
the address computation of the register is done in get_regs function.
Only get the correct set and compute the adress in the caller.

Issue: ALB-10428
Erratum: ERR051700
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>