plat: s32: validate MPIDR input
The TF-A porting guideline specifies that, in case of an invalid MPIDR,
the function 'plat_core_pos_by_mpidr()' should return -1. The function
is executed when the C runtime is available.

Move function 'plat_core_pos_by_mpidr()' to 's32_bl_common.c' and
validate the mpidr input. 'PLATFORM_MAX_CPUS_PER_CLUSTER' is introduced
to validate the CPU ID in a cluster. Use common definitions for the
MPIDR_EL1 register to extract the cluster and CPU IDs and defer from
keeping S32_MPIDR_* definitions which are not necessary.
('S32_MPIDR_CPU_MASK_BITS' is still needed to compute the linear index
of a CPU).

Issue: ALB-10629
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>