ddr: Add dummy function for setting GPRs
If an external DDR Driver/FW is used through the CUSTOM_DDR_DRV
parameter, a dummy implementation for the
`ddrss_gpr_to_io_retention_mode_mmio()` function is needed since
it is not provided by the external DDR code itself. Provide
it via a weak implementation.

This does not affect the flow of the DDR code, since setting the
GPRs would occur in the `ddrss_to_io_retention_mode()` function
provided by the external code, directly through mmio access,
instead of calling a decoupled function to do it.

This leads to incompatibility with SCP-enabled flows, since
the external DDR driver would always directly access the
DDR_GPRs, which would lead to an abort if GPRs are protected
from A53 access.

Issue: ALB-10790
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>